requirements.txt
Django==5.0
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-debug-toolbar==4.2.0
django-import-export==3.3.1
djangorestframework==3.14.0
openpyxl==3.1.2
pillow
python-dotenv==1.0.0